refactor: rename terminusd pkg to webhook/upload
This webhook is a separate service that merely shares its host IP
with terminusd. Rename the package and env var so the name reflects
intent; keep TERMINUSD_HOST as the IP source with a comment explaining
the co-location.

- pkg/integration/terminusd -> pkg/webhook/upload (package upload)
- TERMINUSD_NOTIFY_ENABLED  -> UPLOAD_WEBHOOK_ENABLED
- log prefix [terminusd]    -> [upload-webhook]
- TERMINUSD_HOST stays (only the host portion is borrowed for the
  co-located node IP; port 18832 and path /v1/new_items belong to
  the webhook service, not terminusd).

handlefunc.go imports the new package as alias `uploadwh` because
its own package is also named `upload`.

Made-with: Cursor
